modal components deleted, interchangable modal created, form components added
