+++ src/components/CommentForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./commentForm.component";
