--- src/components/CommentModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./commentModal.component";
