+++ src/components/PostForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./postForm.component";
