--- src/components/PostModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./postModal.component";
