--- src/components/UserModal/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./userModal.component";
